+++ src/MultiLayerProxy/Algorithms/NodeCount.cs
@@ -16,14 +16,13 @@ namespace MultiLayerProxy.Algorithms {
         MultiLayerServer.MessagePassingExtension.GetNodeCountServer(server);
       }
 
-
       List<List<long>> phaseResults =  Proxy.WaitForPhaseResultsAsLong(Phases.NodeCount);
-
       long[] nodeCount = new long[phaseResults[0].Count];
 
+      // Sum up the results from all the servers.
       foreach(List<long> result in phaseResults) {
         for (int i = 0; i < result.Count; i++) {
-            nodeCount[i] += (int) result[i];
+            nodeCount[i] += result[i];
         }
       }
